First pass at implementing the game logic.
This commit implements the core Game of Life logic. The core logic lives
in the GameOfLifeBoard class which is a representation of a 2D board
containing various cells each in one of two states: live or dead.
GameOfLifeBoard includes a Tick() function that implements the rules
described in the Wiki page:
https://en.wikipedia.org/wiki/Conway%27s_Game_of_Life#Rules. The screen
functions and main program driver saw updates to integrate with the new
GameOfLifeBoard object.